Enable top bar in all but lockscreen
