+++ applications/launcher/main.qml
@@ -14,6 +14,7 @@ OxideWindow {
     headerBackgroundColor: "white"
     color: "black"
     backgroundColor: "white"
+    reserveSystemSpace: true
     FontLoader { id: iconFont; source: "/font/icomoon.ttf" }
     onAfterSynchronizing: {
         if (stateController.state == "loading") {
@@ -81,152 +82,6 @@ OxideWindow {
                     onTriggered: {
                         controller.breadcrumb("menu.options", "click", "ui");
                         stateController.state = "settings";
-                    }
-                }
-            }
-        },
-        OxideStatusIcon {
-            source: "qrc:/img/notifications/black.png"
-            text: controller.notificationText
-            visible: controller.hasNotification
-            clip: true
-            Layout.maximumWidth: 300
-            MouseArea {
-                anchors.fill: parent
-                enabled: parent.visible
-                onClicked: {
-                    controller.breadcrumb("notifications", "click", "ui");
-                    stateController.state = "notifications";
-                }
-            }
-        }
-    ]
-    centerMenu: [
-        Label {
-            objectName: "clock"
-            Layout.alignment: Qt.AlignCenter
-            color: window.color
-            MouseArea {
-                anchors.fill: parent
-                onClicked: {
-                    controller.breadcrumb("clock", "click", "ui");
-                    stateController.state = "calendar";
-                }
-            }
-        }
-    ]
-    rightMenu: [
-        OxideStatusIcon {
-            id: wifiState
-            objectName: "wifiState"
-            property string state: "unknown"
-            property int rssi: 0
-            property bool connected: false
-            source: {
-                var icon;
-                if(state === "unknown"){
-                    icon = "unknown";
-                }else if(state === "down"){
-                    icon = "down";
-                }else if(!connected){
-                    icon = "disconnected";
-                }else if(rssi > -50) {
-                    icon = "4_bar";
-                }else if(rssi > -60){
-                    icon = "3_bar";
-                }else if(rssi > -70){
-                    icon = "2_bar";
-                }else if(rssi > -80){
-                    icon = "1_bar";
-                }else{
-                    icon = "0_bar";
-                }
-                return "qrc:/codes.eeems.oxide/img/wifi/" + icon + ".png";
-            }
-            text: controller.showWifiDb ? rssi + "dBm" : ""
-            MouseArea {
-                anchors.fill: parent
-                onClicked: {
-                    controller.breadcrumb("wifi", "click", "ui");
-                    stateController.state = "wifi";
-                }
-            }
-        },
-        OxideStatusIcon {
-            id: batteryLevel
-            objectName: "batteryLevel"
-            property bool alert: false
-            property bool warning: false
-            property bool charging: false
-            property bool connected: false
-            property bool present: true
-            property int level: 0
-            property int temperature: 0
-            source: {
-                var icon = "";
-                if(alert || !present){
-                    icon = "alert";
-                }else if(warning){
-                    icon = "unknown";
-                }else{
-                    if(charging || connected){
-                        icon = "charging_";
-                    }
-                    if(level < 25){
-                        icon += "20";
-                    }else if(level < 35){
-                        icon += "30";
-                    }else if(level < 55){
-                        icon += "50";
-                    }else if(level < 65){
-                        icon += "60";
-                    }else if(level < 85){
-                        icon += "80";
-                    }else if(level < 95){
-                        icon += "90";
-                    }else{
-                        icon += 100;
-                    }
-                }
-                return "qrc:/codes.eeems.oxide/img/battery/" + icon + ".png";
-            }
-            text: (controller.showBatteryPercent ? level + "% " : "") + (controller.showBatteryTemperature ? temperature + "C" : "")
-        },
-        CustomMenu {
-            OxideMenu {
-                id: powerMenu
-                title: qsTr("");
-                font.family: iconFont.name
-                width: 260
-                Action {
-                    text: qsTr(" Suspend")
-                    enabled: !controller.sleepInhibited
-                    onTriggered: {
-                        controller.breadcrumb("menu.suspend", "click", "ui");
-                        controller.suspend();
-                    }
-                }
-                Action {
-                    text: qsTr(" Reboot")
-                    enabled: !controller.powerOffInhibited
-                    onTriggered: {
-                        controller.breadcrumb("menu.reboot", "click", "ui");
-                        controller.reboot();
-                    }
-                }
-                Action {
-                    text: qsTr(" Shutdown")
-                    enabled: !controller.powerOffInhibited
-                    onTriggered: {
-                        controller.breadcrumb("menu.shutdown", "click", "ui");
-                        controller.powerOff();
-                    }
-                }
-                Action {
-                    text: qsTr(" Lock")
-                    onTriggered: {
-                        controller.breadcrumb("menu.lock", "click", "ui");
-                        controller.lock();
                     }
                 }
             }
