+++ applications/process-manager/codes.eeems.erode.oxide
@@ -3,5 +3,5 @@
     "description": "List and kill running processes",
     "icon": "oxide:erode-48",
     "bin": "/home/root/.vellum/bin/erode",
-    "flags": ["nopreload"]
+    "flags": ["nopreload", "topbar"]
 }
