+++ applications/screenshot-viewer/screenshot-viewer.pro
@@ -17,7 +17,7 @@ include(../../qmake/common.pri)
 target.path = $$BIN_INSTALL_PATH
 INSTALLS += target
 
-applications.files = ../../assets/opt/usr/share/applications/codes.eeems.anxiety.oxide
+applications.files = codes.eeems.anxiety.oxide
 applications.path = $$APPLICATIONS_INSTALL_PATH
 INSTALLS += applications
