+++ applications/task-switcher/task-switcher.pro
@@ -18,7 +18,7 @@ include(../../qmake/common.pri)
 target.path = $$BIN_INSTALL_PATH
 INSTALLS += target
 
-applications.files = ../../assets/opt/usr/share/applications/codes.eeems.corrupt.oxide
+applications.files = codes.eeems.corrupt.oxide
 applications.path = $$APPLICATIONS_INSTALL_PATH
 INSTALLS += applications
