+++ applications/topbar/codes.eeems.stain.oxide
@@ -1,8 +1,8 @@
 {
     "displayName": "topbar",
-    "description": "Application launcher",
+    "description": "System top bar",
     "bin": "/home/root/.vellum/bin/stain",
     "type": "background",
-    "flags": ["hidden", "nopreload"],
+    "flags": ["hidden", "nopreload", "noannounce", "privileged", "autoStart"],
     "permissions": ["apps", "wifi", "power", "system", "notification"]
 }
