+++ applications/topbar/controller.h
@@ -3,11 +3,7 @@
 #include <liboxide.h>
 #include <liboxide/sysobject.h>
 
-#include <QGuiApplication>
-#include <QJsonObject>
-#include <QMutableListIterator>
 #include <QObject>
-#include <QTimer>
 
 #include "notificationlist.h"
 #include "wifinetworklist.h"
@@ -21,7 +17,6 @@ using codes::eeems::oxide1::Power;
 using Oxide::SysObject;
 using namespace Oxide::Sentry;
 
-enum State { Normal, PowerSaving };
 enum BatteryState {
   BatteryUnknown,
   BatteryCharging,
@@ -41,10 +36,10 @@ enum WifiState {
   WifiOffline,
   WifiOnline
 };
-enum SwipeDirection { None, Right, Left, Up, Down };
 
 class Controller : public QObject {
   Q_OBJECT
+  Q_PROPERTY(bool visible MEMBER m_visible NOTIFY visibleChanged)
   Q_PROPERTY(bool showWifiDb MEMBER m_showWifiDb NOTIFY showWifiDbChanged)
   Q_PROPERTY(
     bool showBatteryPercent MEMBER m_showBatteryPercent NOTIFY
@@ -76,187 +71,7 @@ class Controller : public QObject {
   )
 public:
   QObject* root = nullptr;
-  explicit Controller(QObject* parent = 0)
-    : QObject(parent)
-    , m_wifion(false)
-    , wifi("/sys/class/net/wlan0")
-    , applications() {
-    networks = new WifiNetworkList();
-    notifications = new NotificationList();
-
-    auto bus = QDBusConnection::systemBus();
-    qDebug() << "Waiting for tarnish to start up";
-    while (!bus.interface()->registeredServiceNames().value().contains(
-      OXIDE_SERVICE
-    )) {
-      struct timespec args{
-        .tv_sec = 1,
-        .tv_nsec = 0,
-      },
-        res;
-      nanosleep(&args, &res);
-    }
-    qDebug() << "Requesting APIs";
-    api = new General(OXIDE_SERVICE, OXIDE_SERVICE_PATH, bus);
-    connect(api, &General::aboutToQuit, qApp, &QGuiApplication::quit);
-    auto reply = api->requestAPI("power");
-    reply.waitForFinished();
-    if (reply.isError()) {
-      qDebug() << reply.error();
-      qFatal("Could not request power API");
-    }
-    auto path = ((QDBusObjectPath)reply).path();
-    if (path == "/") {
-      qDebug() << "API not available";
-      qFatal("Power API was not available");
-    }
-    powerApi = new Power(OXIDE_SERVICE, path, bus);
-    // Connect to signals
-    connect(
-      powerApi,
-      &Power::batteryLevelChanged,
-      this,
-      &Controller::batteryLevelChanged
-    );
-    connect(
-      powerApi,
-      &Power::batteryStateChanged,
-      this,
-      &Controller::batteryStateChanged
-    );
-    connect(
-      powerApi,
-      &Power::batteryTemperatureChanged,
-      this,
-      &Controller::batteryTemperatureChanged
-    );
-    connect(
-      powerApi,
-      &Power::chargerStateChanged,
-      this,
-      &Controller::chargerStateChanged
-    );
-    connect(
-      powerApi, &Power::stateChanged, this, &Controller::powerStateChanged
-    );
-    connect(powerApi, &Power::batteryAlert, this, &Controller::batteryAlert);
-    connect(
-      powerApi, &Power::batteryWarning, this, &Controller::batteryWarning
-    );
-    connect(
-      powerApi, &Power::chargerWarning, this, &Controller::chargerWarning
-    );
-    reply = api->requestAPI("wifi");
-    reply.waitForFinished();
-    if (reply.isError()) {
-      qDebug() << reply.error();
-      qFatal("Could not request wifi API");
-    }
-    path = ((QDBusObjectPath)reply).path();
-    if (path == "/") {
-      qDebug() << "API not available";
-      qFatal("Wifi API was not available");
-    }
-    wifiApi = new Wifi(OXIDE_SERVICE, path, bus);
-    connect(wifiApi, &Wifi::disconnected, this, &Controller::disconnected);
-    connect(
-      wifiApi, &Wifi::networkConnected, this, &Controller::networkConnected
-    );
-    connect(wifiApi, &Wifi::stateChanged, this, &Controller::wifiStateChanged);
-    connect(wifiApi, &Wifi::rssiChanged, this, &Controller::wifiRssiChanged);
-    networks->setAPI(wifiApi);
-    auto state = wifiApi->state();
-    m_wifion = state != WifiState::WifiOff && state != WifiState::WifiUnknown;
-    QTimer::singleShot(1000, [this]() {
-      while (this->root == nullptr) {
-        qApp->processEvents(QEventLoop::AllEvents, 100);
-      }
-      // Get initial values when UI is ready
-      batteryLevelChanged(powerApi->batteryLevel());
-      batteryStateChanged(powerApi->batteryState());
-      batteryTemperatureChanged(powerApi->batteryTemperature());
-      chargerStateChanged(powerApi->chargerState());
-      powerStateChanged(powerApi->state());
-
-      wifiStateChanged(wifiApi->state());
-      wifiRssiChanged(wifiApi->rssi());
-      emit wifiOnChanged(m_wifion);
-      auto network = wifiApi->network();
-      if (network.path() != "/") {
-        networkConnected(network);
-      }
-    });
-    reply = api->requestAPI("system");
-    reply.waitForFinished();
-    if (reply.isError()) {
-      qDebug() << reply.error();
-      qFatal("Could not request system API");
-    }
-    path = ((QDBusObjectPath)reply).path();
-    if (path == "/") {
-      qDebug() << "API not available";
-      qFatal("System API was not available");
-    }
-    systemApi = new System(OXIDE_SERVICE, path, bus);
-    reply = api->requestAPI("apps");
-    reply.waitForFinished();
-    if (reply.isError()) {
-      qDebug() << reply.error();
-      qFatal("Could not request apps API");
-    }
-    path = ((QDBusObjectPath)reply).path();
-    if (path == "/") {
-      qDebug() << "API not available";
-      qFatal("Apps API was not available");
-    }
-    appsApi = new Apps(OXIDE_SERVICE, path, bus);
-    reply = api->requestAPI("notification");
-    reply.waitForFinished();
-    if (reply.isError()) {
-      qDebug() << reply.error();
-      qFatal("Could not request notification API");
-    }
-    path = ((QDBusObjectPath)reply).path();
-    if (path == "/") {
-      qDebug() << "API not available";
-      qFatal("Notification API was not available");
-    }
-    notificationApi = new Notifications(OXIDE_SERVICE, path, bus);
-    connect(
-      notificationApi,
-      &Notifications::notificationAdded,
-      this,
-      &Controller::notificationAdded
-    );
-    connect(
-      notificationApi,
-      &Notifications::notificationRemoved,
-      this,
-      &Controller::notificationRemoved
-    );
-    connect(
-      notificationApi,
-      &Notifications::notificationChanged,
-      this,
-      &Controller::notificationChanged
-    );
-    connect(
-      notifications,
-      &NotificationList::updated,
-      this,
-      &Controller::notificationsUpdated
-    );
-    uiTimer = new QTimer(this);
-    uiTimer->setSingleShot(false);
-    uiTimer->setInterval(3 * 1000); // 3 seconds
-    connect(
-      uiTimer,
-      &QTimer::timeout,
-      this,
-      QOverload<>::of(&Controller::updateUIElements)
-    );
-    uiTimer->start();
-  }
+  explicit Controller(QObject* parent = nullptr);
   Q_INVOKABLE void startup() { loadSettings(); }
   Q_INVOKABLE bool turnOnWifi() {
     if (!wifiApi->enable()) {
@@ -382,6 +197,7 @@ public:
   Apps* getAppsApi() { return appsApi; }
   Notifications* getNotificationApi() { return notificationApi; }
 signals:
+  void visibleChanged(bool);
   void showWifiDbChanged(bool);
   void showBatteryPercentChanged(bool);
   void showBatteryTemperatureChanged(bool);
@@ -394,10 +210,8 @@ signals:
   void notificationTextChanged(QString);
   void notificationsChanged(NotificationList*);
 
-public slots:
-  void updateUIElements();
-
 private slots:
+  void checkCurrentApplication(QDBusObjectPath path = QDBusObjectPath());
   void notificationsUpdated() {
     if (notifications->length() > 1) {
       setNotification(
@@ -652,6 +466,7 @@ private slots:
 
 private:
   void checkUITimer();
+  bool m_visible = false;
   bool m_showWifiDb = false;
   bool m_showBatteryPercent = false;
   bool m_showBatteryTemperature = false;
@@ -662,7 +477,6 @@ private:
   int wifiState = WifiUnknown;
   bool m_wifion;
   SysObject wifi;
-  QTimer* uiTimer;
   General* api = nullptr;
   Power* powerApi = nullptr;
   Wifi* wifiApi = nullptr;
